--- a/tests/NSubstitute.Benchmarks/TestTypes/IInterfaceWithMultipleMethods.cs
+++ b/tests/NSubstitute.Benchmarks/TestTypes/IInterfaceWithMultipleMethods.cs
@@ -1,0 +1,9 @@
+﻿namespace NSubstitute.Benchmarks.TestTypes
+{
+    public interface IInterfaceWithMultipleMethods
+    {
+        int MethodWithArg(int arg);
+        int MethodWithRefArg(ref int arg);
+        int MethodWithOutArg(out int arg);
+    }
+}