… BreakingChanges.md
… src/NSubstitute/Core/Arguments/ArgumentSpecificationQueue.cs
… tests/NSubstitute.Acceptance.Specs/ArgumentMatchingCompat.cs