Closes #401.

As we agreed in the #401, I'm changing the `Arg.XXX()` methods signature to return `ref T`. That allows to use neat syntax to configure both regular and by ref arguments:

```c#
Method(Arg.Any()).Returns(42);
Method(ref Arg.Any()).Returns(42);
Method(out Arg.Any()).Returns(42);
```

I modified all the argument matchers, to cover even the most weird scenarios (I mean `ref Arg.Invoke()`) and be consistent.

I don't create the compat library for now, as believe it should be created on demand only. The reasons:
- that should be a standalone package with its own lifecycle, so it should reside in another repo;
- package should reference NSubstitute v4 from NuGet, which is currently not available.

Performance measurements didn't show any real difference.

[BEFORE:](https://github.com/zvirja/NSubstitute/blob/70e444f2772f4905e0099b17655873affd2db3b2/tests/NSubstitute.Benchmarks/ArgumentSpecificationUsageBenchmark.cs)
```c#
                           Method |  Job | Runtime |      Mean |     Error |    StdDev |
--------------------------------- |----- |-------- |----------:|----------:|----------:|
    ConfigureArgumentWithAnyValue |  Clr |     Clr |  7.920 us | 0.1029 us | 0.0912 us |
 ConfigureRefArgumentWithAnyValue |  Clr |     Clr |  8.712 us | 0.0566 us | 0.0473 us |
 ConfigureOutArgumentWithAnyValue |  Clr |     Clr |  8.577 us | 0.0554 us | 0.0491 us |
    ConfigureArgumentWithAnyValue | Core |    Core | 11.174 us | 0.1102 us | 0.0977 us |
 ConfigureRefArgumentWithAnyValue | Core |    Core | 11.556 us | 0.0439 us | 0.0389 us |
 ConfigureOutArgumentWithAnyValue | Core |    Core | 11.520 us | 0.0815 us | 0.0762 us |
```

AFTER:
```
                           Method |  Job | Runtime |      Mean |     Error |    StdDev |
--------------------------------- |----- |-------- |----------:|----------:|----------:|
    ConfigureArgumentWithAnyValue |  Clr |     Clr |  7.928 us | 0.0529 us | 0.0469 us |
 ConfigureRefArgumentWithAnyValue |  Clr |     Clr |  8.505 us | 0.0920 us | 0.0861 us |
 ConfigureOutArgumentWithAnyValue |  Clr |     Clr |  8.639 us | 0.0277 us | 0.0231 us |
    ConfigureArgumentWithAnyValue | Core |    Core | 11.022 us | 0.0964 us | 0.0855 us |
 ConfigureRefArgumentWithAnyValue | Core |    Core | 11.748 us | 0.0931 us | 0.0871 us |
 ConfigureOutArgumentWithAnyValue | Core |    Core | 11.613 us | 0.0591 us | 0.0524 us |
```

@alexandrnikitin @dtchepak Please review and explicitly approve the change 🍷