@dtchepak 2018-05-19
AFAICT this should work fine [for VB .NET](https://docs.microsoft.com/en-us/dotnet/visual-basic/programming-guide/language-features/procedures/ref-return-values) and [for F# 4.1](https://github.com/fsharp/fslang-design/blob/master/FSharp-4.1/FS-1020-byref-returns.md).

@zvirja, @alexandrnikitin : does that seem correct?