@dtchepak 2018-05-19
I'm happy with this code, with the one disclaimer about compatibility that I've sent a PR for. 😂 

@zvirja, if you can live with the compatibility addition, and if @alexandrnikitin is happy with this and `CompatArg`, please merge 👍