@zvirja 2018-05-20
Decided to reply here, so it's easier to track the history. From the tech perspective the suggested change (external PR) is totally fine.

---
@dtchepak You created a PR to add `CompatArg` directly to the `NSubstitute` library. I don't _strongly_ mind to do that, however I don't like it. Let's discuss all the points to ensure that we considered everything 😉 

I would prefer to have `CompatArg` as a separate NuGet library and here is my reasoning:
- This is a temporary workaround, but it's very unclear when we can remove it. In a year or two we might be still in doubt whether we have clients using this. Meaning, we'll need to have this code for the long time as a part of the primary library. As for the `NSubstitute.Compat` package, it might live forever - we don't care about it and don't need to maintain its code as a part of the primary development.
- Our documentation says nothing about `CompatArg` and we use `Arg` everywhere. When `CompatArg` is shipped as a part of the `NSubstitute` library, it makes it a kind of "official" API, which should also be covered in documentation. Of course, we don't want to spread workaround across the whole documentation. In this regards the standalone `NSubstitute.Compat` package is much better, as any stuff present there _is not_ a part of the official API.
- By moving code to the extra-package, we'll encourage people to upgrade the language version, rather than immediately switch to the compat usage. Of course, they could still install the package, but they will better feel that it's a temp workaround and they will tend to remove it.
- If we decide to rely on new language features in some other places, the external compatibility package will be a good known place for the compatibility stuff and shims will not pollute the primary library.
- By having extra-package it will be much simpler to have statistics how wide is the usage. It might help us to easily make the decisions in the future.

I don't see that installation of extra-package is a big deal if users are indeed required to use NSubstitute library in the compatibility mode. We are targeting the test projects, so no blockers there. From other hand, the benefits mentioned above are valuable to me.

@dtchepak @alexandrnikitin Please review my points and let me know whether you still persist on making `CompatArg` as part of the `NSubstitute` library. If so - I'll merge in the suggested change.

Thanks 😉