@dtchepak 2018-05-21
@zvirja @alexandrnikitin I appreciate the discussion! 😄 

> By moving code to the extra-package, we'll encourage people to upgrade the language version, rather than immediately switch to the compat usage.

My main problem with this is that I don't think it is generally the developers objecting to upgrading language version. Most likely they're toiling away in a department that won't spring to upgrade their VS 2013 licences. :-\ I really don't want to make life harder for these people. 

I'm wondering if we should leave the current API as is and expose the new methods in a different way. Then all current code will continue to work the same, and people can opt in. I was thinking of static imports (`using static`) to help with this.

```csharp
public static class ArgMatchers {
    public static ref T Any<T>() { ... }
}

// In new code:
using static NSubstitute.ArgMatchers;
// ...
sub.Foo(Any<int>()).Returns("hi");
```

I think that should avoid any breaking changes. If people are already using static imports of `Arg` it won't cause a problem, and they can update the `using` to point to the new matchers if they want the new functionality. This should ease documentation overhead as well, it will all work depending on the import.

----

Back to the compat stuff, I don't mind it in a new package. How will the package bounds work? When does the package need to be up-issued? If we are using a new package, then I think I prefer `ArgC` to make it easier for people to use (once they've gone through the hoop of installing a new package). Should the code be in a new repo, or just another csproj here that we can build as a separate package?