@zvirja 2018-05-21
> I was thinking of static imports (using static) to help with this.

Unfortunately, that will break NUnit. Just gave it a quick check:
![image](https://user-images.githubusercontent.com/1074182/40301694-8d081d02-5cf5-11e8-9ca3-3f3273b29a2e.png)

It causes ambiguity between `Arg.Is()` and `Is` from NUnit world :-/

What I'm also worried about is that it will cause devs to perform a set of preparation actions for each file - they will quickly start to hate us if that's the API we force them to use.
Additionally, as usual, there might be code style rules preventing `using static` usage (as it's purely about the syntax), so we cannot rely on this feature.

Therefore, probably, we cannot proceed with that way 😕 

----
> Back to the compat stuff, I don't mind it in a new package. 

After the Alex's explanation I thought that it might be indeed an overkill to create a package for the single file. If later we find that we have more compat stuff, I'd suggest to get back to the discussion.

I've merged the https://github.com/zvirja/NSubstitute/pull/5 with minor alternations (moved to `NSubstitute.Compatibility` namespace, tweaked the doc.

@dtchepak If you are fine with the code now, please approve it, so I'll proceed with a merge 😉