@dtchepak 2018-05-21
@zvirja I'm still sad at having to inconvenience my developer brothers and sisters stuck on pre-C#7... but I'll approve 'cause I think we've got a reasonable compromise. 😄 

We'll have to update breakingchanges.md to also tell them to import `NSubstitute.Compatibility`.