@dtchepak 2018-05-21
We could avoid the static import problem by naming the members `ArgIs` and `ArgAny`, but that probably looks a bit confusing.