@zvirja 2018-05-21
> We'll have to update breakingchanges.md to also tell them to import `NSubstitute.Compatibility`

I don't see too much value from that. Nowadays IDEs show you suggestions from non-imported namespaces as well (or suggest you to import them). The `CompatArg` name is quite unique, so they'll quickly figure out the namespace they require to import 😄